Moved helper functions to a seperate file.
